pyproject.toml
[project]
name = "arc-agi-babies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Méo Desbois-Renaudin", email = "meo-externe.desbois-renaudin@edf.fr" }
]
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.5.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyqt5==5.15.10",
]

[project.scripts]
google-agent = "arc_agi_babies:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
